split up examples, add ._read(), expanded docs
