+++ package.json
@@ -22,11 +22,10 @@
   },
   "homepage": "https://github.com/noopkat/azure-iothub-receiver#readme",
   "dependencies": {
-    "azure-event-hubs": "0.0.8",
+    "azure-event-hubs": "0.0.8"
+  },
+  "devDependencies": {
     "dotenv": "^4.0.0",
-    "events": "^1.1.1",
-    "stream": "0.0.2",
-    "through2": "^2.0.3",
-    "util": "^0.10.3"
+    "through2": "^2.0.3"
   }
 }
